requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
python-dotenv==1.0.0
scikit-learn==1.3.2
openai==1.3.5
anthropic==0.7.8
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
dj-database-url==2.1.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-cors-headers==4.3.1
whitenoise==6.6.0
watchfiles==1.1.1

